Changded platoon_len from uint64 to uint8

+++ enclave/lib/Enclave/Enclave.cpp
@@ -13,7 +13,7 @@
 ec256_key_pair_t *key_pair = NULL; // Global EC256 cache
 int position = 0;
 sgx_ec256_public_t *pub_keys = NULL;
-uint64_t platoon_len = 0;
+uint8_t platoon_len = 0;
 ////////////
 // PUBLIC //
 ////////////
@@ -76,11 +76,11 @@ sgx_status_t setPosition(int *pos) {
 // It takes 2 parameters
 // pubkeys : sgx_ec256_public_t pubkeys*
 //	A pointer to an array of pubkeys to be set, ordered by position in
-//platoon
+// platoon
 // platoon_len : uint8_t
 //	The length of platoon
 sgx_status_t setPubKeys(sgx_ec256_public_t *pub_keys_in,
-                        uint64_t platoon_len_in) {
+                        uint8_t platoon_len_in) {
   int retval = 0;
 
   // Free the memory if it has been allocated previously
